+++ src/strigui/label.clj
@@ -37,4 +37,5 @@
   wdg/Widget
   (coord [this canvas] (coord-label this canvas))
   (defaults [this] this)
+  (before-drawing [this] this)
   (draw [this canvas] (draw-label this canvas)))